Initial Design Work
Exploring syntax possibilities for declaring functions, types, and variables. The goal is to have as unified and extensible a syntax as possible that manages to be simple and usable from the programmers perspective. The design of the type system and syntax currently takes a lot of influence from scala at the moment, though the "context-contract" conception is unique to my knowledge.