Mut Changing Rules Dev Work
Adds a lot of work and development on settling down the rules for mutability changing in regards to various situations within the language. Needed to determine whether some pointer ideas are feasible.
Adds a possible new idea for pointers that merges some of the old designs (really just a reworking of an old idea in conjunction with mut rules work)
Improves documentation of the memory work
Updates speroc reference